pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["wonderland/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    # "ARG", # Unused function argument
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    # "NPY",
    # "AIR",
    "PERF",
    # "FURB",
    # "LOG",
    "RUF",
]
ignore = [
    "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
    "SIM102", # sometimes it's better to nest
    "UP038", # Checks for uses of isinstance/issubclass that take a tuple
    # of types for comparison.
    # Deactivated because it can make the code slow:
    # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
  "Werkzeug[watchdog]==3.0.6", # https://github.com/pallets/werkzeug
  "ipdb==0.13.13", # https://github.com/gotcha/ipdb
  "psycopg[c]==3.2.3", # https://github.com/psycopg/psycopg
  "watchfiles==0.24.0", # https://github.com/samuelcolvin/watchfiles
  # Testing
  # ------------------------------------------------------------------------------
  "mypy==1.11.2", # https://github.com/python/mypy
  "django-stubs[compatible-mypy]==5.1.0", # https://github.com/typeddjango/django-stubs
  "pytest==8.3.3", # https://github.com/pytest-dev/pytest
  "pytest-sugar==1.0.0", # https://github.com/Frozenball/pytest-sugar
  "djangorestframework-stubs==3.15.1", # https://github.com/typeddjango/djangorestframework-stubs
  # Documentation
  # ------------------------------------------------------------------------------
  "sphinx==7.4.7", # https://github.com/sphinx-doc/sphinx
  "sphinx-autobuild==2024.10.3", # https://github.com/GaretJax/sphinx-autobuild
  # Code quality
  # ------------------------------------------------------------------------------
  "ruff==0.6.9", # https://github.com/astral-sh/ruff
  "coverage==7.6.1", # https://github.com/nedbat/coveragepy
  "djlint==1.35.2", # https://github.com/Riverside-Healthcare/djLint
  "pre-commit==3.8.0", # https://github.com/pre-commit/pre-commit
  # Django
  # ------------------------------------------------------------------------------
  "factory-boy==3.3.1", # https://github.com/FactoryBoy/factory_boy
  "django-debug-toolbar==4.4.6", # https://github.com/jazzband/django-debug-toolbar
  "django-extensions==3.2.3", # https://github.com/django-extensions/django-extensions
  "django-coverage-plugin==3.1.0", # https://github.com/nedbat/django_coverage_plugin
  "pytest-django==4.9.0", # https://github.com/pytest-dev/pytest-django
]

[project]
name = "wonderland"
version = "0.1.0"
description = "Into the Rabbithole"
readme = "README.md"

license = { text = "MIT" }
authors = [
  { name = "Alexander Vega", email = "ajaradv@gmail.com" },
]
requires-python = "==3.12.*"
dependencies = [
  "python-slugify==8.0.4", # https://github.com/un33k/python-slugify
  "Pillow==11.0.0", # https://github.com/python-pillow/Pillow
  "argon2-cffi==23.1.0", # https://github.com/hynek/argon2_cffi
  "whitenoise==6.7.0", # https://github.com/evansd/whitenoise
  "redis==5.1.1", # https://github.com/redis/redis-py
  "hiredis==3.1.0", # https://github.com/redis/hiredis-py
  "celery==5.4.0", # pyup: < 6.0  # https://github.com/celery/celery
  "django-celery-beat==2.7.0", # https://github.com/celery/django-celery-beat
  "flower==2.0.1", # https://github.com/mher/flower
  "uvicorn[standard]==0.34.0", # https://github.com/encode/uvicorn
  "uvicorn-worker==0.2.0", # https://github.com/Kludex/uvicorn-worker
  # Django
  # ------------------------------------------------------------------------------
  "django==5.0.9", # pyup: < 5.1  # https://www.djangoproject.com/
  "django-environ==0.11.2", # https://github.com/joke2k/django-environ
  "django-model-utils==5.0.0", # https://github.com/jazzband/django-model-utils
  "django-allauth[mfa]==65.0.2", # https://github.com/pennersr/django-allauth
  "django-crispy-forms==2.3", # https://github.com/django-crispy-forms/django-crispy-forms
  "crispy-bootstrap5==2024.10", # https://github.com/django-crispy-forms/crispy-bootstrap5
  "django-redis==5.4.0", # https://github.com/jazzband/django-redis
  # Django REST Framework
  "djangorestframework==3.15.2", # https://github.com/encode/django-rest-framework
  "django-cors-headers==4.4.0", # https://github.com/adamchainz/django-cors-headers
  # DRF-spectacular for api documentation
  "drf-spectacular==0.27.2", # https://github.com/tfranzel/drf-spectacular
  "django-webpack-loader==3.1.1", # https://github.com/django-webpack/django-webpack-loader
  "gunicorn==23.0.0", # https://github.com/benoitc/gunicorn
  "psycopg[c]==3.2.3", # https://github.com/psycopg/psycopg
  "sentry-sdk==2.19.2", # https://github.com/getsentry/sentry-python
  # Django
  # ------------------------------------------------------------------------------
  "django-storages[s3]==1.14.4", # https://github.com/jschneier/django-storages
  "django-anymail[mailjet]==12.0", # https://github.com/anymail/django-anymail
  "django-htmx>=1.21.0",
  "django-tailwind[reload]>=3.8.0",
  "crispy-tailwind>=1.0.3",
  "fontawesomefree>=6.6.0",
  "channels>=4.2.0",
  "channels-redis>=4.2.1",
  "django-template-partials>=24.4",
  "django-cotton>=1.5.2",
  "django-unfold>=0.44.0",
  "django-import-export>=4.3.3",
  "django-simple-history>=3.7.0",
  "django-allauth-ui>=1.5.3",
  "django-widget-tweaks>=1.5.0",
  "slippers>=0.6.2",
]
